Removed ESLint Mantine config and Header component with associated styles and tests.
