--- app/components/Header.module.css
@@ -1,57 +1,0 @@
-.header {
-  height: rem(70px);
-}
-
-.link {
-  display: block;
-  line-height: 1;
-  padding: rem(8px) rem(12px);
-  border-radius: var(--mantine-radius-sm);
-  text-decoration: none;
-  font-size: var(--mantine-font-size-md);
-  font-weight: 500;
-}
-
-.linkLight {
-  color: light-dark(var(--mantine-color-gray-7), var(--mantine-color-dark-0));
-
-  @mixin hover {
-    background-color: light-dark(
-      var(--mantine-color-gray-0),
-      var(--mantine-color-dark-6)
-    );
-  }
-
-  [data-mantine-color-scheme] &[data-active] {
-    background-color: #fff;
-    color: #000;
-  }
-}
-
-.linkDark {
-  color: white;
-
-  [data-mantine-color-scheme] &[data-active] {
-    color: white;
-  }
-}
-
-.startYourSuccess {
-  font-weight: 600;
-  background-color: #8a60f6;
-  color: white;
-  border-radius: var(--mantine-radius-lg);
-
-  @mixin hover {
-    background-color: #8a60f6;
-  }
-
-  [data-mantine-color-scheme] &[data-active] {
-    background-color: #8a60f6;
-    color: white;
-  }
-}
-
-.linkLabel {
-  margin-right: rem(5px);
-}
