--- app/components/Layout.tsx
@@ -1,129 +1,0 @@
-import {Await, useLocation} from '@remix-run/react';
-import {Suspense} from 'react';
-import type {
-  CartApiQueryFragment,
-  FooterQuery,
-  HeaderQuery,
-} from 'storefrontapi.generated';
-import {Aside} from '~/components/Aside';
-import {CartMain} from '~/components/Cart';
-import {Footer} from '~/components/Footer';
-import {Header, HeaderMenu} from '~/components/Header';
-import {
-  PredictiveSearchForm,
-  PredictiveSearchResults,
-} from '~/components/Search';
-
-export type LayoutProps = {
-  cart: Promise<CartApiQueryFragment | null>;
-  children?: React.ReactNode;
-  footer: Promise<FooterQuery>;
-  header: HeaderQuery;
-  isLoggedIn: Promise<boolean>;
-};
-
-export function Layout({
-  cart,
-  children = null,
-  footer,
-  header,
-  isLoggedIn,
-}: LayoutProps) {
-  const location = useLocation();
-  const path = location.pathname;
-
-  return (
-    <>
-      {!path.includes('/account') ? (
-        <>
-          <CartAside cart={cart} />
-          {!path.includes('/artist/') ? (
-            <>
-              <SearchAside />
-              <MobileMenuAside menu={header.menu} shop={header.shop} />
-              <Header header={header} cart={cart} isLoggedIn={isLoggedIn} />
-            </>
-          ) : null}
-        </>
-      ) : null}
-      <main>{children}</main>
-      {!path.includes('/account') && !path.includes('/artist/') ? (
-        <Suspense>
-          <Await resolve={footer}>
-            {(footer) => <Footer menu={footer.menu} shop={header.shop} />}
-          </Await>
-        </Suspense>
-      ) : null}
-    </>
-  );
-}
-
-function CartAside({cart}: {cart: LayoutProps['cart']}) {
-  return (
-    <Aside id="cart-aside" heading="Indkøbskurv">
-      <Suspense fallback={<p>Indlæser indkøbskurv ...</p>}>
-        <Await resolve={cart}>
-          {(cart) => {
-            return <CartMain cart={cart} layout="aside" />;
-          }}
-        </Await>
-      </Suspense>
-    </Aside>
-  );
-}
-
-function SearchAside() {
-  return (
-    <Aside id="search-aside" heading="SEARCH">
-      <div className="predictive-search">
-        <br />
-        <PredictiveSearchForm>
-          {({fetchResults, inputRef}) => (
-            <div>
-              <input
-                name="q"
-                onChange={fetchResults}
-                onFocus={fetchResults}
-                placeholder="Search"
-                ref={inputRef}
-                type="search"
-              />
-              &nbsp;
-              <button
-                onClick={() => {
-                  window.location.href = inputRef?.current?.value
-                    ? `/search?q=${inputRef.current.value}`
-                    : `/search`;
-                }}
-              >
-                Search
-              </button>
-            </div>
-          )}
-        </PredictiveSearchForm>
-        <PredictiveSearchResults />
-      </div>
-    </Aside>
-  );
-}
-
-function MobileMenuAside({
-  menu,
-  shop,
-}: {
-  menu: HeaderQuery['menu'];
-  shop: HeaderQuery['shop'];
-}) {
-  return (
-    menu &&
-    shop?.primaryDomain?.url && (
-      <Aside id="mobile-menu-aside" heading="MENU">
-        <HeaderMenu
-          menu={menu}
-          viewport="mobile"
-          primaryDomainUrl={shop.primaryDomain.url}
-        />
-      </Aside>
-    )
-  );
-}
