+++ app/root.tsx
@@ -27,6 +27,7 @@ import {
   useWindowScroll,
 } from '@mantine/hooks';
 import {
+  Await,
   isRouteErrorResponse,
   Link,
   Links,
@@ -36,6 +37,7 @@ import {
   Scripts,
   ScrollRestoration,
   useLoaderData,
+  useLocation,
   useMatches,
   useRouteError,
   type ShouldRevalidateFunction,
@@ -46,9 +48,15 @@ import {
   type LoaderFunctionArgs,
   type SerializeFrom,
 } from '@shopify/remix-oxygen';
-import {IconActivity, IconChevronRight} from '@tabler/icons-react';
-import {Layout} from '~/components/Layout';
+import {
+  IconChevronRight,
+  IconDashboard,
+  IconLogin,
+  IconShoppingCartPlus,
+} from '@tabler/icons-react';
+import {Suspense} from 'react';
 import favicon from './assets/favicon.svg';
+import {Footer} from './components/Footer';
 import appStyles from './styles/app.css?url';
 import logo from '/Artboard4.svg';
 
@@ -142,6 +150,8 @@ export default function App() {
   const pinned = useHeadroom({fixedAt: 120});
   const [scroll] = useWindowScroll();
   const isMobile = useMediaQuery('(max-width: 62em)');
+  const location = useLocation();
+  const path = location.pathname;
 
   return (
     <html lang="en">
@@ -158,123 +168,204 @@ export default function App() {
       <body>
         <MantineProvider>
           <ModalsProvider>
-            <AppShell
-              header={{
-                height: 70,
-                collapsed: !pinned && !isMobile,
-                offset: opened,
-              }}
-              navbar={{
-                width: 300,
-                breakpoint: 'sm',
-                collapsed: {desktop: true, mobile: !opened},
-              }}
-              withBorder={false}
-            >
-              <AppShell.Header
-                component={Flex}
-                p="md"
-                bg={scroll.y === 0 ? 'transparent' : 'white'}
+            {!path.includes('/account') && !path.includes('/artist/') ? (
+              <AppShell
+                header={{
+                  height: 70,
+                  collapsed: !pinned && !isMobile,
+                  offset: opened,
+                }}
+                navbar={{
+                  width: 300,
+                  breakpoint: 'sm',
+                  collapsed: {desktop: true, mobile: !opened},
+                }}
+                withBorder={false}
               >
-                <Group h="100%" miw="150px">
-                  <Burger
-                    opened={opened}
-                    onClick={toggle}
-                    hiddenFrom="sm"
-                    size="sm"
-                  />
-                  <UnstyledButton component={Link} to="/">
-                    <Title
-                      order={1}
-                      component={Flex}
-                      lh="xs"
-                      fz={rem(20)}
-                      fw="500"
-                      data-testid="logo-login"
-                    >
-                      ByS
-                      <Image
-                        src={logo}
-                        alt="it's me"
-                        h="auto"
-                        w="8px"
-                        mx="2px"
-                      />
-                      sters
-                    </Title>
-                  </UnstyledButton>
-                </Group>
-                <Flex
-                  justify="center"
-                  align="center"
-                  visibleFrom="sm"
-                  style={{flex: 1}}
-                  h="100%"
+                <AppShell.Header
+                  component={Flex}
+                  p="md"
+                  bg={scroll.y === 0 ? 'transparent' : 'white'}
                 >
-                  {data.header.menu?.items.map(({url, title}) => (
-                    <NavLink key={url} to={new URL(url || '').pathname}>
-                      {({isActive}) => (
-                        <Button
-                          radius="xl"
-                          fz="md"
-                          fw="500"
-                          variant={
-                            title.includes('skønhedskarriere')
-                              ? 'filled'
-                              : 'transparent'
-                          }
-                          color={
-                            title.includes('skønhedskarriere')
-                              ? '#8a60f6'
-                              : 'black'
-                          }
-                        >
-                          {title}
-                        </Button>
-                      )}
-                    </NavLink>
-                  ))}
-                </Flex>
-                <Flex
-                  justify="flex-end"
-                  align="center"
-                  gap={0}
-                  visibleFrom="sm"
-                  miw="150px"
-                >
-                  login
-                </Flex>
-              </AppShell.Header>
-
-              <AppShell.Navbar px={4}>
-                {data.header.menu?.items.map(({url, title}) => (
-                  <NavLink
-                    key={url}
-                    to={new URL(url || '').pathname}
-                    onClick={toggle}
+                  <Group h="100%" miw="150px">
+                    <Burger
+                      opened={opened}
+                      onClick={toggle}
+                      hiddenFrom="sm"
+                      size="sm"
+                    />
+                    <UnstyledButton component={Link} to="/">
+                      <Title
+                        order={1}
+                        component={Flex}
+                        lh="xs"
+                        fz={rem(20)}
+                        fw="500"
+                        data-testid="logo-login"
+                      >
+                        ByS
+                        <Image
+                          src={logo}
+                          alt="it's me"
+                          h="auto"
+                          w="8px"
+                          mx="2px"
+                        />
+                        sters
+                      </Title>
+                    </UnstyledButton>
+                  </Group>
+                  <Flex
+                    justify="center"
+                    align="center"
+                    visibleFrom="sm"
+                    style={{flex: 1}}
+                    h="100%"
                   >
-                    {({isActive}) => (
-                      <NavLinkMantine
-                        label={title}
-                        leftSection={<IconActivity size="1rem" stroke={1.5} />}
-                        rightSection={
-                          <IconChevronRight
-                            size="0.8rem"
-                            stroke={1.5}
-                            className="mantine-rotate-rtl"
-                          />
+                    {data.header.menu?.items.map(({url, title}) => (
+                      <NavLink key={url} to={new URL(url || '').pathname}>
+                        {({isActive}) => (
+                          <Button
+                            radius="xl"
+                            fz="md"
+                            fw="500"
+                            variant={
+                              title.includes('skønhedskarriere')
+                                ? 'filled'
+                                : 'transparent'
+                            }
+                            color={
+                              title.includes('skønhedskarriere')
+                                ? '#8a60f6'
+                                : 'black'
+                            }
+                          >
+                            {title}
+                          </Button>
+                        )}
+                      </NavLink>
+                    ))}
+                  </Flex>
+                  <Flex
+                    justify="flex-end"
+                    align="center"
+                    gap={0}
+                    visibleFrom="sm"
+                    miw="150px"
+                  >
+                    <Suspense>
+                      <Await resolve={data.isLoggedIn}>
+                        {(isLoggedIn) => (
+                          <Button
+                            variant="outline"
+                            size="compact-md"
+                            color="black"
+                            component={Link}
+                            to="/account"
+                            rightSection={
+                              isLoggedIn ? <IconDashboard /> : <IconLogin />
+                            }
+                          >
+                            {isLoggedIn ? 'Dashboard' : 'Log ind'}
+                          </Button>
+                        )}
+                      </Await>
+                    </Suspense>
+                    <Suspense>
+                      <Await resolve={data.cart}>
+                        {(cart) =>
+                          cart?.totalQuantity && cart?.totalQuantity > 0 ? (
+                            <Button
+                              variant="outline"
+                              size="compact-md"
+                              color="black"
+                              component={Link}
+                              to="/cart"
+                              rightSection={<IconShoppingCartPlus />}
+                            >
+                              Indkøbskurv {cart.totalQuantity}
+                            </Button>
+                          ) : null
                         }
-                        active={isActive}
-                      />
-                    )}
-                  </NavLink>
-                ))}
-              </AppShell.Navbar>
+                      </Await>
+                    </Suspense>
+                  </Flex>
+                </AppShell.Header>
 
-              <AppShell.Main>
-                <Outlet />
-              </AppShell.Main>
-            </AppShell>
+                <AppShell.Navbar px={4}>
+                  {data.header.menu?.items.map(({url, title}) => (
+                    <NavLinkMantine
+                      component={NavLink}
+                      key={url}
+                      to={new URL(url || '').pathname}
+                      onClick={toggle}
+                      label={title}
+                      rightSection={
+                        <IconChevronRight
+                          size="0.8rem"
+                          stroke={1.5}
+                          className="mantine-rotate-rtl"
+                        />
+                      }
+                    />
+                  ))}
+                  <Flex gap="xs" p="xs">
+                    <Suspense>
+                      <Await resolve={data.isLoggedIn}>
+                        {(isLoggedIn) => (
+                          <Button
+                            variant="outline"
+                            size="compact-lg"
+                            color="black"
+                            component={Link}
+                            to="/account"
+                            rightSection={
+                              isLoggedIn ? <IconDashboard /> : <IconLogin />
+                            }
+                            fullWidth
+                          >
+                            {isLoggedIn ? 'Dashboard' : 'Log ind'}
+                          </Button>
+                        )}
+                      </Await>
+                    </Suspense>
+                    <Suspense>
+                      <Await resolve={data.cart}>
+                        {(cart) =>
+                          cart?.totalQuantity && cart?.totalQuantity > 0 ? (
+                            <Button
+                              variant="outline"
+                              size="compact-lg"
+                              color="black"
+                              component={Link}
+                              to="/cart"
+                              onClick={toggle}
+                              rightSection={<IconShoppingCartPlus />}
+                              fullWidth
+                            >
+                              Indkøbskurv {cart.totalQuantity}
+                            </Button>
+                          ) : null
+                        }
+                      </Await>
+                    </Suspense>
+                  </Flex>
+                </AppShell.Navbar>
+
+                <AppShell.Main>
+                  <Outlet />
+                  <Suspense>
+                    <Await resolve={data.footer}>
+                      {(footer) => (
+                        <Footer menu={footer.menu} shop={data.header.shop} />
+                      )}
+                    </Await>
+                  </Suspense>
+                </AppShell.Main>
+              </AppShell>
+            ) : (
+              <Outlet />
+            )}
           </ModalsProvider>
         </MantineProvider>
         <ScrollRestoration nonce={nonce} />
@@ -314,17 +405,15 @@ export function ErrorBoundary() {
       <body>
         <MantineProvider theme={theme}>
           <ModalsProvider>
-            <Layout {...rootData}>
-              <div className="route-error">
-                <h1>Oops</h1>
-                <h2>{errorStatus}</h2>
-                {errorMessage && (
-                  <fieldset>
-                    <pre>{errorMessage}</pre>
-                  </fieldset>
-                )}
-              </div>
-            </Layout>
+            <div className="route-error">
+              <h1>Oops</h1>
+              <h2>{errorStatus}</h2>
+              {errorMessage && (
+                <fieldset>
+                  <pre>{errorMessage}</pre>
+                </fieldset>
+              )}
+            </div>
           </ModalsProvider>
         </MantineProvider>
         <ScrollRestoration nonce={nonce} />
